requirements.txt
streamlit
pandas
folium
streamlit-folium
plotly
openpyxl
gspread
google-auth